Sound System implementation
Very basic implementation of a sound system (using irrKlang library
(free for non-commercial use)).

An example of a playing sound and music in the TestEnvironment should
give you a rough idea on how to use it.

Also, this system doesn't make differentiation between sounds & musics,
so i got rid of the music element of the resourcesholder since we won't
be using them anymore.

Keep in mind that this is a first implementation and it is in no way
flawless

In case of dll errors, you'll find them in the directory:
(src\Libs\irrKlang-1.5.0\bin\win32-visualStudio)

I also tried a new standard for the descriptive comments of functions